Extract SharedSystem to separate module
